remove DistanceCalculator
- No caching is preferable to caching, especially when not even the
  most basic performance benchmarks exist or are required

+++ include/common_typedefs.h
@@ -6,7 +6,7 @@
 namespace MoleculeManip {
 
 /* Common typedefs */
-enum class BondType : uint8_t {
+enum class BondType : unsigned {
     Single,
     Double,
     Triple,
